backend/requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.22.0
SQLAlchemy==2.0.20
psycopg2-binary==2.9.6
python-dotenv==1.0.0
passlib[bcrypt]==1.7.5
PyJWT==2.8.0
pydantic==2.4.1
alembic==1.12.0
httpx==0.26.0